Move the stuff related to runners to lambda_shared
