+++ tests/ci/ci_runners_metrics_lambda/lambda_shared
@@ -1,0 +1,1 @@
+../lambda_shared_package/lambda_shared
