+++ tests/ci/lambda_shared_package/lambda_shared/__init__.py
@@ -4,7 +4,8 @@ It exists as __init__.py and lambda_shared/__init__.py to work both in local and
 import json
 import logging
 import time
-from typing import List, Optional
+from collections import namedtuple
+from typing import Any, List, Optional
 
 import boto3  # type: ignore
 import requests  # type: ignore
@@ -21,10 +22,14 @@ RUNNER_TYPE_LABELS = [
 
 
 ### VENDORING
-def get_parameter_from_ssm(name, decrypt=True, client=None):
+def get_parameter_from_ssm(
+    name: str, decrypt: bool = True, client: Optional[Any] = None
+) -> str:
     if not client:
         client = boto3.client("ssm", region_name="us-east-1")
-    return client.get_parameter(Name=name, WithDecryption=decrypt)["Parameter"]["Value"]
+    return client.get_parameter(Name=name, WithDecryption=decrypt)[  # type: ignore
+        "Parameter"
+    ]["Value"]
 
 
 class CHException(Exception):
@@ -65,10 +70,64 @@ class ClickHouseHelper:
 
         raise CHException("Cannot fetch data from clickhouse")
 
-    def select_json_each_row(self, db: str, query: str) -> List[dict]:  # type: ignore
+    def select_json_each_row(self, db: str, query: str) -> List[dict]:
         text = self._select_and_get_json_each_row(db, query)
         result = []
         for line in text.split("\n"):
             if line:
                 result.append(json.loads(line))
         return result
+
+
+### Runners
+
+RunnerDescription = namedtuple(
+    "RunnerDescription", ["id", "name", "tags", "offline", "busy"]
+)
+RunnerDescriptions = List[RunnerDescription]
+
+
+def list_runners(access_token: str) -> RunnerDescriptions:
+    headers = {
+        "Authorization": f"token {access_token}",
+        "Accept": "application/vnd.github.v3+json",
+    }
+    per_page = 100
+    response = requests.get(
+        f"https://api.github.com/orgs/ClickHouse/actions/runners?per_page={per_page}",
+        headers=headers,
+    )
+    response.raise_for_status()
+    data = response.json()
+    total_runners = data["total_count"]
+    print("Expected total runners", total_runners)
+    runners = data["runners"]
+
+    # round to 0 for 0, 1 for 1..100, but to 2 for 101..200
+    total_pages = (total_runners - 1) // per_page + 1
+
+    print("Total pages", total_pages)
+    for i in range(2, total_pages + 1):
+        response = requests.get(
+            "https://api.github.com/orgs/ClickHouse/actions/runners"
+            f"?page={i}&per_page={per_page}",
+            headers=headers,
+        )
+        response.raise_for_status()
+        data = response.json()
+        runners += data["runners"]
+
+    print("Total runners", len(runners))
+    result = []
+    for runner in runners:
+        tags = [tag["name"] for tag in runner["labels"]]
+        desc = RunnerDescription(
+            id=runner["id"],
+            name=runner["name"],
+            tags=tags,
+            offline=runner["status"] == "offline",
+            busy=runner["busy"],
+        )
+        result.append(desc)
+
+    return result
