+++ tests/ci/terminate_runner_lambda/lambda_shared
@@ -1,0 +1,1 @@
+../lambda_shared_package/lambda_shared
