+++ tests/ci/terminate_runner_lambda/requirements.txt
@@ -1,3 +1,3 @@
-requests<2.30
+../lambda_shared_package
 PyJWT
 cryptography<38
